Modified mapred.go
Added rpc implementation to check for hosts willing to work. Modified
structure of worker. Workers can now only be specifie as a slice and
not as a network address. This is because we expect the host to running
the client machine and therefore the user is expected to know these
locations beforehand.